update the stub_request
We updated the hardcoded alegre:3100 to use Settings.get_safe_url,
and removed the Host option from Headers. When trying to use the
Settings.get_safe_url for the Host, we would get an error, because it
expects it not to have the http. Since for this test the Host isn't
important we decided to remove it.