CV2-2772 add safe url acessing (#344)
* CV2-2772 add safe url accessing

* change alegre.local to alegre:3100

This way it works. But I'm not sure why.
Making this commit to save this state, but still working on it.

* update the stub_request

We updated the hardcoded alegre:3100 to use Settings.get_safe_url,
and removed the Host option from Headers. When trying to use the
Settings.get_safe_url for the Host, we would get an error, because it
expects it not to have the http. Since for this test the Host isn't
important we decided to remove it.

* updated the stub_request for pender_client_test

Same issue as the alegre_client, we were getting an error
because of the Host. Since it's not necessary we removed it.
Also updated to use Settings.get_safe_url instead of hardcoding
the url.

* add platform to docker-compose

I wasn't able to build on the mac m1, adding platform fixed that.