update the way we remove ' - THIP Media'
The way we were doing this before was causing a few empty strings for
the headlines. We also added raw_claim_review["raw_response"]["title"]["rendered"]
as a fallback, in case some of the titles are truly missing.

I looked into using the ld_json, but this meant parsing the url twice,
so we didn't go with it.